Updates to allow test-spy feature to be used with XCTest in Xcode 5.
The new source files and headers are now included in the Kiwi-XCTest target, and one new compiler error, related to passing NSUInteger arguments (which vary in size depending upon the architecture) to vararg methods like printf, had to be dealt with (by forcing the arguments to be long ints).

+++ Classes/Matchers/KWHaveReceivedMatcher.m
@@ -205,10 +205,10 @@
     NSUInteger actualArgumentCount = [argumentMatchers count];
     if (actualArgumentCount < messageArgumentCount) {
         [NSException raise:NSInvalidArgumentException
-                    format:@"%@ message takes %u argument(s), but only %u argument matcher(s) given",
+                    format:@"%@ message takes %lu argument(s), but only %lu argument matcher(s) given",
                     NSStringFromSelector(aSelector),
-                    messageArgumentCount,
-                    actualArgumentCount];
+                    (unsigned long)messageArgumentCount,
+                    (unsigned long)actualArgumentCount];
     }
     KWMessagePattern *messagePattern =
         [KWMessagePattern messagePatternWithSelector:aSelector
